Its Alive!
The program now works and simulates correctly, with the PlayPause button
wired up to control the simulation thread. To do this, a new class
"ThreadHandler" was made. Also some comments/documentation has been
added to better explain the functions of the program. Still more to do,
but the program is functional.