New Feature Added!
Now on start up, a dialog will appear prompting the user to enter in the
length and width of the grid. The maximum size is 50x50 cells because
anything larger causes the program to lag severely. I'm sure 2500
buttons will be enough for most applications. Addittionally, the main
method has been moved from MainWindow to the new class StartWindow to
handle this change.