New Feature 2 Added!
Now there is a slider at the top of the MainWindow that lets you control
the speed of the simulation. The speed can cause the program to update
as fast as once every 50ms or as slow as once every second. Also, I have
increased the limit of buttons from 50x50 to 100x100 because with the
ability to adjust the speed, you can get the program to function fine
with 100 buttons if you set the speed slower. I wouldn't recomend
running it full speed with 100x100 buttons however.